Move metrics to separate module
