+++ core/lib/snapshots_applier/Cargo.toml
@@ -1,14 +1,13 @@
 [package]
 name = "zksync_snapshots_applier"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
 authors = ["The Matter Labs Team <hello@matterlabs.dev>"]
 homepage = "https://zksync.io/"
 repository = "https://github.com/matter-labs/zksync-era"
 license = "MIT OR Apache-2.0"
 keywords = ["blockchain", "zksync"]
 categories = ["cryptography"]
-
 
 [dependencies]
 zksync_dal = { path = "../../lib/dal" }
